feat(lualine): add function as indicator support
This commit introduces the ability to use a function as an indicator
in the lualine setup. The function will receive a harpoon_entry as an
argument and should return a string to be used as the indicator. This
provides more flexibility for users to customize their lualine
configuration.

A check is added to determine if the indicator is a function or a
string before adding it to the status. If it's a function, it's
called with the harpoon_entry as an argument. If it's a string, it's
added directly to the status.

This change is backward compatible as it doesn't affect the existing
usage of string indicators.